+++ gridtools/__init__.py
@@ -1,6 +1,6 @@
 
 __all__ = ["app", "bathyutils", "fileutils", "datasource", "gridutils",
-        "meshrefinement", "spherical", "sysinfo", "topoutils"]
+        "meshrefinement", "spherical", "sysinfo", "topoutils", "utils"]
 
 # Copied from sphinx/src/sphinx/__init__.py
 # Remove warnings support