+++ gridtools/topoutils.py
@@ -37,6 +37,8 @@ class TopoUtils:
         finer resolution than the grid file. It is also assumed that the
         topography values are defined at the cell centers.
 
+        This routine is based on code from :cite:p:`Simkins_2021_xesmf_regridding`.
+
         We recommend using
         `GEBCO2020 topography dataset <https://www.gebco.net/data_and_products/gridded_bathymetry_data/>`_
         for the topography file.
@@ -45,34 +47,34 @@ class TopoUtils:
           gridFile: Path to netCDF gridFile        ==> grd: grid class object
           topoFile: Path to netCDF topography file ==> dsName: data source name
 
-        Keyword Arguments:
-          gridGeoLoc: "center" or "corner" - cell location of geographic
-            placement of grid file - for example, x/y are located at 'q'
-            points, which are located at the corner of each cell.
-          topoVarName: variable name of the topography variable within
-            the topography file. It is assumed that the variable represents
-            elevation. If it is not representing elevation, please set
-            'convert_to_depth' to False.
-          coarsenInt: Integer value used to decrease resolution of a
-            given topography file - see `xarray.coarsen`
-          superGrid: When true, this assumes the gridFile is a supergrid
-            and the resulting topography is coarsened to a regular grid.  Not
-            currently implemented.
-          method: Regrid method of xesmf regridder. Options are 'conservative',
-            'bilinear', 'nearests2d', 'nearestd2s', 'patch' - options can be
-            read about here
-            https://pangeo-xesmf.readthedocs.io/en/latest/notebooks/Compare_algorithms.html
-          periodic: Boolean, either True or False - When dealing with global
-            grids, we need to set periodic=True, otherwise data along the
-            meridian line will be missing.
-          gridDimX: The name of the dimension along the X axis of the grid file
-          gridDimY: The name of the dimension along the Y axis of the grid file
-          gridLatName: The name of the latitude variable of the grid file
-          gridLonName: The name of the longitude variable of the grid file
-          topoDimX: The name of the dimension along the X axis of the topography file
-          topoDimY: The name of the dimension along the Y axis of the topography file
-          topoLatName: The name of the latitude variable within the topography file
-          topoLonName: The name of the longitude variable within the topography file
+        **Keyword arguments**:
+
+            * gridGeoLoc: "center" or "corner" - cell location of geographic
+              placement of grid file - for example, x/y are located at 'q'
+              points, which are located at the corner of each cell.
+            * topoVarName: variable name of the topography variable within
+              the topography file. It is assumed that the variable represents
+              elevation. If it is not representing elevation, please set
+              'convert_to_depth' to False.
+            * coarsenInt: Integer value used to decrease resolution of a
+              given topography file - see `xarray.coarsen`
+            * superGrid: When true, this assumes the gridFile is a supergrid
+              and the resulting topography is coarsened to a regular grid.  Not
+              currently implemented.
+            * method: Regrid method of xesmf regridder. Options are 'conservative',
+              'bilinear', 'nearests2d', 'nearestd2s' or 'patch'.  See:
+              `pangeo-xesmf regridding options <https://pangeo-xesmf.readthedocs.io/en/latest/notebooks/Compare_algorithms.html>`_.
+            * periodic: Boolean, either True or False - When dealing with global
+              grids, we need to set periodic=True, otherwise data along the
+              meridian line will be missing.
+            * gridDimX: The name of the dimension along the X axis of the grid file
+            * gridDimY: The name of the dimension along the Y axis of the grid file
+            * gridLatName: The name of the latitude variable of the grid file
+            * gridLonName: The name of the longitude variable of the grid file
+            * topoDimX: The name of the dimension along the X axis of the topography file
+            * topoDimY: The name of the dimension along the Y axis of the topography file
+            * topoLatName: The name of the latitude variable within the topography file
+            * topoLonName: The name of the longitude variable within the topography file
 
         """